requirements.txt
inject
rx

awscli
boto3

luigi
